Copy the core of WP_SQLite_Translator to WP_SQLite_Driver
Additionally, move the current SQLite driver prototype to a temporary class
WP_SQLite_Driver_Prototype to be gradually moved to the new driver.
